Remove support for pyOpenSSL SSLContext
