+++ docs/reference/hip.contrib.rst
@@ -4,14 +4,6 @@ urllib3.contrib package
 These modules implement various extra features, that may not be ready for
 prime time or that require optional third-party dependencies.
 
-urllib3.contrib.pyopenssl module
---------------------------------
-
-.. automodule:: hip.contrib.pyopenssl
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 urllib3.contrib.socks module
 ----------------------------
