+++ setup.cfg
@@ -46,12 +46,6 @@ where = src
 
 [options.extras_require]
 brotli = brotlipy>=0.6.0
-secure =
-    pyOpenSSL>=0.14
-    cryptography>=1.3.4
-    idna>=2.0.0
-    certifi
-    ipaddress; python_version=='2.7'
 socks = PySocks >=1.5.6, <2.0, !=1.5.7
 
 [tool:pytest]
